Repo // Implementing set() for NSWindowController.
